[Fixs #109] Adding confirmation on delete actions

+++ commands/droplets.go
@@ -67,8 +67,9 @@ func Droplet() *Command {
 
 	AddStringSliceFlag(cmdDropletCreate, doctl.ArgVolumeList, []string{}, "Volumes to attach")
 
-	CmdBuilder(cmd, RunDropletDelete, "delete ID [ID|Name ...]", "Delete droplet by id or name", Writer,
+	cmdRunDropletDelete := CmdBuilder(cmd, RunDropletDelete, "delete ID [ID|Name ...]", "Delete droplet by id or name", Writer,
 		aliasOpt("d", "del", "rm"), docCategories("droplet"))
+	AddBoolFlag(cmdRunDropletDelete, doctl.ArgDeleteForce, false, "Force droplet delete")
 
 	CmdBuilder(cmd, RunDropletGet, "get", "get droplet", Writer,
 		aliasOpt("g"), displayerType(&droplet{}), docCategories("droplet"))
@@ -267,7 +268,7 @@ func RunDropletCreate(c *CmdConfig) error {
 
 	wg.Wait()
 	close(errs)
-	
+
 	item := &droplet{droplets: createdList}
 	c.Display(item)
 
@@ -424,6 +425,11 @@ func allInt(in []string) ([]int, error) {
 func RunDropletDelete(c *CmdConfig) error {
 	ds := c.Droplets()
 
+	force, err := c.Doit.GetBool(c.NS, doctl.ArgDeleteForce)
+	if err != nil {
+		return err
+	}
+
 	tagName, err := c.Doit.GetString(c.NS, doctl.ArgTagName)
 	if err != nil {
 		return err
@@ -434,20 +440,31 @@ func RunDropletDelete(c *CmdConfig) error {
 	} else if len(c.Args) > 0 && tagName != "" {
 		return fmt.Errorf("please specify droplets identifiers or a tag name")
 	} else if tagName != "" {
-		return ds.DeleteByTag(tagName)
-	}
-
-	fn := func(ids []int) error {
-		for _, id := range ids {
-			if err := ds.Delete(id); err != nil {
-				return fmt.Errorf("unable to delete droplet %d: %v", id, err)
-			}
+		if force || AskForConfirm("delete droplet by \""+tagName+"\" tag") == nil {
+			return ds.DeleteByTag(tagName)
+		} else {
+			return fmt.Errorf("Operation aborted.")
 		}
-
 		return nil
 	}
 
-	return matchDroplets(c.Args, ds, fn)
+	if force || AskForConfirm("delete droplet(s)") == nil {
+
+		fn := func(ids []int) error {
+			for _, id := range ids {
+				if err := ds.Delete(id); err != nil {
+					return fmt.Errorf("unable to delete droplet %d: %v", id, err)
+				}
+			}
+			return nil
+		}
+		return matchDroplets(c.Args, ds, fn)
+	} else {
+		return fmt.Errorf("Operation aborted.")
+	}
+
+	return nil
+
 }
 
 type matchDropletsFn func(ids []int) error
